+++ .vscode/settings.json
@@ -7,6 +7,5 @@
     "sonarlint.connectedMode.project": {
         "connectionId": "SonarQube for IDE - Visual Studio Code",
         "projectKey": "mhagnumdw_rewrite-format-sql"
-    },
-    "java.compile.nullAnalysis.mode": "automatic"
+    }
 }
